+++ px-oci-mon/Dockerfile
@@ -1,6 +1,6 @@
 FROM alpine:latest
 
 RUN apk add --update util-linux
-ADD px-runcds /
-ENTRYPOINT [ "/px-runcds" ]
+ADD px-oci-mon /
+ENTRYPOINT [ "/px-oci-mon" ]
