+++ px-oci-mon/main.go
@@ -8,7 +8,7 @@ import (
 	"syscall"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/portworx/px-installer/px-runcds/utils"
+	"github.com/portworx/px-installer/px-oci-mon/utils"
 )
 
 const (
