+++ px-oci-mon/utils/inspect.go
